# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Convert a GitHub Issue/Pull Request/Discussion URL into Markdown.",
        "title": "issue2md Web API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/convert": {
            "post": {
                "description": "Fetch one GitHub issue, pull request, or discussion and render it as markdown.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "convert"
                ],
                "summary": "Convert GitHub URL to Markdown",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub issue/pull/discussion URL",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "markdown body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "resource not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "rate limited",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "render failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "upstream failure",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/openapi.json": {
            "get": {
                "description": "Returns generated OpenAPI JSON. Run `make swagger` before calling this endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docs"
                ],
                "summary": "Get OpenAPI specification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "read failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "spec unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}